Fix web integration tests.
