+++ packages/datadog_session_replay/test/processor/processor_worker_test.dart
@@ -19,6 +19,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
+import '../test_utils.dart';
+
 class MockDatadogSessionReplayPlatform extends Mock
     with MockPlatformInterfaceMixin
     implements DatadogSessionReplayPlatform {}
